Move TypedMessenger
