+++ src/util/mod.js
@@ -10,6 +10,6 @@ export {toFormattedJsonString} from "./toFormattedJsonString.js";
 export * from "./util.js";
 export * from "./bufferUtil.js";
 export * from "./meshShapeGeneration/mod.js";
-export * from "./TypedMessenger.js";
+export * from "./TypedMessenger/TypedMessenger.js";
 export * as ErrorSerialization from "./TypedMessenger/errorSerialization.js";
 export * from "./WeakValueMap.js";
