+++ test/unit/studio/src/tasks/shared/basicWorker.js
@@ -1,4 +1,4 @@
-import {TypedMessenger} from "../../../../../../src/util/TypedMessenger.js";
+import {TypedMessenger} from "../../../../../../src/util/TypedMessenger/TypedMessenger.js";
 
 /** @typedef {typeof responseHandlers} BasicWorkerResponseHandlers */
