+++ test/unit/studio/src/windowManagement/contentWindows/ContentWindowBuildView/ContentWindowBuildView.test.js
@@ -2,7 +2,7 @@ import {getMockArgs} from "../shared.js";
 import {ContentWindowBuildView} from "../../../../../../../studio/src/windowManagement/contentWindows/ContentWindowBuildView/ContentWindowBuildView.js";
 import {runWithDomAsync} from "../../../../shared/runWithDom.js";
 import {GestureInProgressManager} from "../../../../../../../studio/src/misc/GestureInProgressManager.js";
-import {TypedMessenger} from "../../../../../../../src/util/TypedMessenger.js";
+import {TypedMessenger} from "../../../../../../../src/util/TypedMessenger/TypedMessenger.js";
 import {assertEquals} from "std/testing/asserts.ts";
 
 /**
